better error tracking and everything better --- structured dot graphs (knowing which dissectors)

+++ src/preprocessors/Profinet/dissect/buffer/Buffy-int.h
@@ -39,25 +39,25 @@ struct Buffy_ops {
    * @param this the calling Buffer
    * @param bitOffset the bit offset the new buffer starts at
    */
-  Buffy_t *(*Buffy_createVirtual)(Buffy_t *this, unsigned int bitOffset);
+  Buffy_t *(*Buffy_createVirtual)(Buffy_t *this, int bitOffset);
 
-  uint8_t *(*Buffy_copyNBytes)(Buffy_t *this, uint8_t *dest, unsigned int byteCount, unsigned int byteOffset);
+  uint8_t *(*Buffy_copyNBytes)(Buffy_t *this, uint8_t *dest, int byteCount, int byteOffset);
 
-  uint8_t (*Buffy_getBits8)(Buffy_t *this, unsigned int byteOffset);
+  uint8_t (*Buffy_getBits8)(Buffy_t *this, int byteOffset);
 
-  uint16_t (*Buffy_getBits16)(Buffy_t *this, unsigned int byteOffset);
+  uint16_t (*Buffy_getBits16)(Buffy_t *this, int byteOffset);
 
-  uint32_t (*Buffy_getBits32)(Buffy_t *this, unsigned int byteOffset);
+  uint32_t (*Buffy_getBits32)(Buffy_t *this, int byteOffset);
 
-  uint64_t (*Buffy_getBits64)(Buffy_t *this, unsigned int byteOffset);
+  uint64_t (*Buffy_getBits64)(Buffy_t *this, int byteOffset);
 
-  uint8_t (*Buffy_getBitsWalk8)(Buffy_t *this, unsigned int *byteOffset);
+  uint8_t (*Buffy_getBitsWalk8)(Buffy_t *this, int *byteOffset);
 
-  uint16_t (*Buffy_getBitsWalk16)(Buffy_t *this, unsigned int *byteOffset);
+  uint16_t (*Buffy_getBitsWalk16)(Buffy_t *this, int *byteOffset);
 
-  uint32_t (*Buffy_getBitsWalk32)(Buffy_t *this, unsigned int *byteOffset);
+  uint32_t (*Buffy_getBitsWalk32)(Buffy_t *this, int *byteOffset);
 
-  uint64_t (*Buffy_getBitsWalk64)(Buffy_t *this, unsigned int *byteOffset);
+  uint64_t (*Buffy_getBitsWalk64)(Buffy_t *this, int *byteOffset);
 
   /**
    * @brief Get 1 - 8 bits returned in a uint8.
@@ -68,7 +68,7 @@ struct Buffy_ops {
    *
    * @return unsigned 8 bit value representing the specified bit range
    */
-  uint8_t (*Buffy_getNBits8)(Buffy_t *this, unsigned int bitOffset,
+  uint8_t (*Buffy_getNBits8)(Buffy_t *this, int bitOffset,
       const int noOfBits);
 
   /**
@@ -80,7 +80,7 @@ struct Buffy_ops {
    *
    * @return unsigned 16 bit value representing the specified bit range
    */
-  uint16_t (*Buffy_getNBits16)(Buffy_t *this, unsigned int bitOffset,
+  uint16_t (*Buffy_getNBits16)(Buffy_t *this, int bitOffset,
       const int noOfBits, const enum Encoding encoding);
 
   /**
@@ -92,7 +92,7 @@ struct Buffy_ops {
    Gu
    * @return unsigned 32 bit value representing the specified bit range
    */
-  uint32_t (*Buffy_getNBits32)(Buffy_t *this, unsigned int bitOffset,
+  uint32_t (*Buffy_getNBits32)(Buffy_t *this, int bitOffset,
       const int noOfBits, const enum Encoding encoding);
 
   /**
@@ -104,7 +104,7 @@ struct Buffy_ops {
    *
    * @return unsigned 64 bit value representing the specified bit range
    */
-  uint64_t (*Buffy_getNBits64)(Buffy_t *this, unsigned int bitOffset,
+  uint64_t (*Buffy_getNBits64)(Buffy_t *this, int bitOffset,
       const int noOfBits, const enum Encoding encoding);
 
 	/**
@@ -136,7 +136,7 @@ struct Buffy {
 	struct Buffy *next;
 
 	/** The position in the raw packet data of this buffer without any offset **/
-	uint64_t position;
+	int position;
 
 	/** Whether this buffer was initialized. **/
 	bool initialized;
